[STYLE] common 패키지 분리
