+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/auth/controller/dto/request/MemberRequestDTO.java
@@ -1,0 +1,7 @@
+package com.nonsoolmate.nonsoolmateServer.domain.auth.controller.dto.request;
+
+import com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums.PlatformType;
+import jakarta.validation.constraints.NotNull;
+
+public record MemberRequestDTO(@NotNull PlatformType platformType) {
+}
