+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/auth/exception/AuthException.java
@@ -1,0 +1,10 @@
+package com.nonsoolmate.nonsoolmateServer.domain.auth.exception;
+
+import com.nonsoolmate.nonsoolmateServer.global.error.exception.BusinessException;
+import com.nonsoolmate.nonsoolmateServer.global.error.exception.BusinessExceptionType;
+
+public class AuthException extends BusinessException {
+    public AuthException(BusinessExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
