+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/auth/service/vo/MemberSignUpVO.java
@@ -1,0 +1,25 @@
+package com.nonsoolmate.nonsoolmateServer.domain.auth.service.vo;
+
+import com.nonsoolmate.nonsoolmateServer.domain.member.entity.Member;
+import com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums.PlatformType;
+import com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums.Role;
+import com.nonsoolmate.nonsoolmateServer.external.oauth.service.vo.enums.AuthType;
+import lombok.Builder;
+
+@Builder
+public record MemberSignUpVO(Long memberId, String email, String name, PlatformType platformType, Role role,
+                             String birthYear, String gender, String phoneNumber, AuthType authType) {
+    public static MemberSignUpVO of(Member member, PlatformType platformType, AuthType authtype) {
+        return MemberSignUpVO.builder()
+                .memberId(member.getMemberId())
+                .email(member.getEmail())
+                .name(member.getName())
+                .platformType(platformType)
+                .role(Role.ROLE_USER)
+                .birthYear(member.getBirthYear())
+                .gender(member.getGender())
+                .phoneNumber(member.getPhoneNumber())
+                .authType(authtype)
+                .build();
+    }
+}
