+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/external/aws/service/dummy.txt
@@ -1,0 +1,1 @@
+프로젝트 진행 시 해당 파일 삭제하시면 됩니다
