+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/external/oauth/service/vo/enums/AuthType.java
@@ -1,0 +1,5 @@
+package com.nonsoolmate.nonsoolmateServer.external.oauth.service.vo.enums;
+
+public enum AuthType {
+    SIGN_UP, LOGIN, SIGN_OUT
+}
