--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/auth/controller/dto/request/MemberRequestDTO.java
@@ -1,7 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.auth.controller.dto.request;
-
-import com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums.PlatformType;
-import jakarta.validation.constraints.NotNull;
-
-public record MemberRequestDTO(@NotNull PlatformType platformType) {
-}
