--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/auth/exception/AuthException.java
@@ -1,10 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.auth.exception;
-
-import com.nonsoolmate.nonsoolmateServer.global.error.exception.BusinessException;
-import com.nonsoolmate.nonsoolmateServer.global.error.exception.BusinessExceptionType;
-
-public class AuthException extends BusinessException {
-    public AuthException(BusinessExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
