--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/auth/service/vo/enums/AuthType.java
@@ -1,5 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.auth.service.vo.enums;
-
-public enum AuthType {
-    SIGN_UP, LOGIN, SIGN_OUT
-}
