--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/config/SwaggerConfig.java
@@ -1,13 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.config;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.info.Info;
-import org.springframework.context.annotation.Configuration;
-
-@OpenAPIDefinition(
-        info = @Info(title = "NONSOOLMATE",
-                description = "NONSOOLMATE SERVER API Documentation",
-                version = "v1"))
-@Configuration
-public class SwaggerConfig {
-}
